fix: resolved additional feedback
